Changing aggregate version to long and Segregating Order infrastructure layers

+++ src/Services/Order/Application/Abstractions/EventSourcing/EventStore/Events/StoreEvent.cs
@@ -7,7 +7,7 @@ public abstract record StoreEvent<TAggregate, TId>
     where TAggregate : IAggregateRoot<TId>
     where TId : struct
 {
-    public int Version { get; }
+    public long Version { get; }
     public TId AggregateId { get; init; }
     public string AggregateName { get; } = typeof(TAggregate).Name;
     public string EventName { get; init; }
